[wip] TensorConversions: bypass allocating new data storage when converting from unified memory to CPU.
[Important: This is heave wip and non-functional]

Idea here is to avoid allocating a new data storage when we're moving a
tensor from device to CPU when using unified memory.  It is faster in that
case to simply dereference the memory.

This code is still WiP and currently suffers from an issue where a new
tensor isn't returned, so metadata aren't updated and as a result, the
following snippet won't work:

  device_tensor = torch.rand(1, device='cuda')
  assert device_tensor.to(device='cpu') == torch.device('cpu')

The code needs to be placed at a lower level, when a new tensor has been
allocated but its Storage hasn't.

Signed-off-by: Quentin Casasnovas <quentin.casasnovas@flex.ai>